Move types
Hide private types.
Move types to files actually using them.
Add const and avoid pointless references.
