+++ src/scenario.hpp
@@ -34,6 +34,9 @@ struct ScenarioOptions {
     std::vector<std::string> disabledExtensions{};
 };
 
+struct DispatchComputeData;
+struct DispatchDataGraphData;
+
 class Scenario {
   public:
     /// \brief Constructor
